+++ vllm/model_executor/layers/fused_moe/layer.py
@@ -38,11 +38,8 @@ from vllm.model_executor.layers.fused_moe.rocm_aiter_fused_moe import (
 from vllm.model_executor.layers.fused_moe.router.router_factory import (
     create_fused_moe_router,
 )
-from vllm.model_executor.layers.fused_moe.runner.moe_runner import (
-    MoERunner,
-)
-from vllm.model_executor.layers.fused_moe.runner.moe_runner_interface import (
-    MoERunnerInterface,
+from vllm.model_executor.layers.fused_moe.runner.moe_runner_factory import (
+    create_moe_runner,
 )
 from vllm.model_executor.layers.fused_moe.runner.shared_experts import (
     SharedExperts,
@@ -589,7 +586,7 @@ class FusedMoE(PluggableLayer):
         # Storing the runner in the FusedMoE is an intermediate state, eventually
         # the runner will own the FusedMoE layer and provide the execution interface
         # for MoE ops.
-        self.runner: MoERunnerInterface = MoERunner(
+        self.runner = create_moe_runner(
             layer_name=self.layer_name,
             moe_config=self.moe_config,
             router=self.router,
